+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/activites/DashboardActivity.java
@@ -1,8 +1,12 @@
 package com.mh.systems.demoapp.activites;
 
+import android.content.BroadcastReceiver;
+import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
+import android.preference.PreferenceManager;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -14,22 +18,27 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.GoogleApiAvailability;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
 import com.mh.systems.demoapp.R;
 import com.mh.systems.demoapp.adapter.RecyclerAdapter.DashboardRecyclerAdapter;
 import com.mh.systems.demoapp.constants.ApplicationGlobal;
-import com.mh.systems.demoapp.web.WebAPI;
 import com.mh.systems.demoapp.models.DeleteToken.AJsonParamsDeleteToken;
 import com.mh.systems.demoapp.models.DeleteToken.DeleteTokenAPI;
 import com.mh.systems.demoapp.models.DeleteToken.DeleteTokenResult;
 import com.mh.systems.demoapp.models.UnreadNewsCount.AJsonParamsGetUnreadCount;
 import com.mh.systems.demoapp.models.UnreadNewsCount.GetUnreadNewsCountAPI;
 import com.mh.systems.demoapp.models.UnreadNewsCount.GetUnreadNewsResponse;
-import com.mh.systems.demoapp.models.UnreadNewsCount.UnreadNewsCountData;
+import com.mh.systems.demoapp.models.featuresflag.AJsonParamsFeaturesFlag;
+import com.mh.systems.demoapp.models.featuresflag.FeatureFlagsAPI;
+import com.mh.systems.demoapp.models.featuresflag.FeatureFlagsResponse;
 import com.mh.systems.demoapp.models.weather.WeatherApiResponse;
 import com.mh.systems.demoapp.push.QuickstartPreferences;
+import com.mh.systems.demoapp.push.RegistrationIntentService;
+import com.mh.systems.demoapp.web.WebAPI;
 import com.mh.systems.demoapp.web.api.WebServiceMethods;
 
 import java.lang.reflect.Type;
@@ -45,7 +54,7 @@ import retrofit.RetrofitError;
 
 /**
  * The {@link DashboardActivity} used to display {@link GridView}, Settings and
- * Logout option. Basically, it will be use as the ForecastMain screen of application
+ * Logout option. Basically, it will be use as the welcome screen of application
  * after Login.
  *
  * @author {@link karan@ucreate.co.in}
@@ -53,7 +62,9 @@ import retrofit.RetrofitError;
  */
 public class DashboardActivity extends BaseActivity {
 
-    private final String LOG_TAG = DashboardActivity.class.getSimpleName();
+    private String LOG_TAG = DashboardActivity.class.getSimpleName();
+
+    private final int PLAY_SERVICES_RESOLUTION_REQUEST = 9000;
 
     /*********************************
      * INSTANCES OF CLASSES
@@ -109,7 +120,17 @@ public class DashboardActivity extends BaseActivity {
     AJsonParamsGetUnreadCount aJsonParamsGetUnreadCount;
 
     GetUnreadNewsResponse getUnreadNewsResponse;
-    UnreadNewsCountData unreadNewsCountData;
+
+    /**
+     * Instances of Features Flag
+     * on dashboard.
+     */
+    FeatureFlagsAPI featureFlagsAPI;
+    AJsonParamsFeaturesFlag aJsonParamsFeaturesFlag;
+
+    FeatureFlagsResponse featureFlagsResponse;
+
+    BroadcastReceiver mRegistrationBroadcastReceiver;
 
     /*********************************
      * INSTANCES OF LOCAL DATA TYPE
@@ -148,7 +169,14 @@ public class DashboardActivity extends BaseActivity {
          */
         ButterKnife.bind(DashboardActivity.this);
 
-        setGridMenuOptions();
+        //Initialize adapter.
+        dashboardRecyclerAdapter = new DashboardRecyclerAdapter(this,
+                dashboardItemsArrayList,
+                iHandicapPosition,
+                loadPreferenceValue(ApplicationGlobal.KEY_HCAP_EXACT_STR, "N/A"));
+        gvMenuOptions.setAdapter(dashboardRecyclerAdapter);
+
+        sendTokenToServer();
 
         //LogOut listener.
         llLogoutBtn.setOnClickListener(mLogoutListener);
@@ -202,12 +230,34 @@ public class DashboardActivity extends BaseActivity {
 //        }
     }
 
+    /**
+     * Implements this method to send Token to
+     * server for push notifications.
+     */
+    private void sendTokenToServer() {
+        mRegistrationBroadcastReceiver = new BroadcastReceiver() {
+            @Override
+            public void onReceive(Context context, Intent intent) {
+                //  mRegistrationProgressBar.setVisibility(ProgressBar.GONE);
+                SharedPreferences sharedPreferences =
+                        PreferenceManager.getDefaultSharedPreferences(context);
+                boolean sentToken = sharedPreferences
+                        .getBoolean(QuickstartPreferences.SENT_TOKEN_TO_SERVER, false);
+            }
+        };
+
+        if (checkPlayServices()) {
+            Intent intent = new Intent(this, RegistrationIntentService.class);
+            startService(intent);
+        }
+    }
+
     @Override
     protected void onResume() {
         super.onResume();
 
-        //Call Weather api functionality.
         if (isOnline(DashboardActivity.this)) {
+            callFeaturesFlagService();
             callWeatherService();
             getUnreadNewsCountService();
         }
@@ -218,6 +268,9 @@ public class DashboardActivity extends BaseActivity {
      * dynamically.
      */
     private void setGridMenuOptions() {
+
+        dashboardItemsArrayList.clear();
+        iHandicapPosition = -1;
 
         //Add Handicap.
         if (loadPreferenceBooleanValue(ApplicationGlobal.KEY_HANDICAP_FEATURE, false)) {
@@ -273,8 +326,7 @@ public class DashboardActivity extends BaseActivity {
         }
 
         //Set Grid options adapter.
-        dashboardRecyclerAdapter = new DashboardRecyclerAdapter(this, dashboardItemsArrayList, iHandicapPosition, loadPreferenceValue(ApplicationGlobal.KEY_HCAP_EXACT_STR, "N/A"));
-        gvMenuOptions.setAdapter(dashboardRecyclerAdapter);
+        dashboardRecyclerAdapter.notifyDataSetChanged();
 
         setupGridLayout(dashboardItemsArrayList.size());
 
@@ -286,12 +338,11 @@ public class DashboardActivity extends BaseActivity {
      * Grid.
      */
     private void setupGridLayout(int iGridSize) {
-        // Create a grid layout with two columns
+
         GridLayoutManager layoutManager = new GridLayoutManager(this, 6);
 
         switch (iGridSize) {
             case 3:
-                // Create a custom SpanSizeLookup where the first item spans both columns
                 layoutManager.setSpanSizeLookup(new GridLayoutManager.SpanSizeLookup() {
                     @Override
                     public int getSpanSize(int position) {
@@ -301,7 +352,6 @@ public class DashboardActivity extends BaseActivity {
                 break;
 
             case 4:
-                // Create a custom SpanSizeLookup where the first item spans both columns
                 layoutManager.setSpanSizeLookup(new GridLayoutManager.SpanSizeLookup() {
                     @Override
                     public int getSpanSize(int position) {
@@ -311,7 +361,6 @@ public class DashboardActivity extends BaseActivity {
                 break;
 
             case 5:
-                // Create a custom SpanSizeLookup where the first item spans both columns
                 layoutManager.setSpanSizeLookup(new GridLayoutManager.SpanSizeLookup() {
                     @Override
                     public int getSpanSize(int position) {
@@ -321,7 +370,6 @@ public class DashboardActivity extends BaseActivity {
                 break;
 
             default:
-                // Create a custom SpanSizeLookup where the first item spans both columns
                 layoutManager.setSpanSizeLookup(new GridLayoutManager.SpanSizeLookup() {
                     @Override
                     public int getSpanSize(int position) {
@@ -461,7 +509,6 @@ public class DashboardActivity extends BaseActivity {
 
                     @Override
                     public void failure(RetrofitError error) {
-                        //you can handle the errors here
                         Log.e(LOG_TAG, "RetrofitError : " + error);
                     }
                 });
@@ -469,7 +516,7 @@ public class DashboardActivity extends BaseActivity {
 
     private void updateSuccessResponse(JsonObject jsonObject) {
 
-        Log.e(LOG_TAG, "Weather response : " + jsonObject.toString());
+        Log.e(LOG_TAG, "WEATHER RESPONSE : " + jsonObject.toString());
 
         Type type = new TypeToken<WeatherApiResponse>() {
         }.getType();
@@ -498,9 +545,7 @@ public class DashboardActivity extends BaseActivity {
 //            savePreferenceValue(ApplicationGlobal.KEY_TEMPKEY_IMAGE, ("e"+weatherData.getWeather().get(0).getIcon()));
 
         } else {
-            //  Toast.makeText(DashboardActivity.this, weatherApiResponse.getMessage(), Toast.LENGTH_LONG).show();
             Log.e(LOG_TAG, weatherApiResponse.getMessage());
-            //callWeatherService();
         }
     }
 
@@ -536,15 +581,13 @@ public class DashboardActivity extends BaseActivity {
             @Override
             public void success(JsonObject jsonObject, retrofit.client.Response response) {
 
-                updateTokenSuccessResponse(jsonObject);
+                deleteTokenSuccessResponse(jsonObject);
             }
 
             @Override
             public void failure(RetrofitError error) {
-                //you can handle the errors here
                 Log.e(LOG_TAG, "RetrofitError : " + error);
                 hideProgress();
-
                 showAlertMessage("" + getResources().getString(R.string.error_please_retry));
             }
         });
@@ -554,9 +597,9 @@ public class DashboardActivity extends BaseActivity {
      * Implements a method which called when token
      * deleted from web server successfully.
      */
-    private void updateTokenSuccessResponse(JsonObject jsonObject) {
+    private void deleteTokenSuccessResponse(JsonObject jsonObject) {
 
-        Log.e(LOG_TAG, "SUCCESS RESULT : " + jsonObject.toString());
+        Log.e(LOG_TAG, "DELETE TOKEN RESPONSE : " + jsonObject.toString());
 
         Type type = new com.newrelic.com.google.gson.reflect.TypeToken<DeleteTokenResult>() {
         }.getType();
@@ -575,27 +618,24 @@ public class DashboardActivity extends BaseActivity {
                 startActivity(new Intent(DashboardActivity.this, LoginActivity.class));
                 finish();
             } else {
-                /*mAwesomeValidation.addValidation(etUserName, "regex", updatePasswordResponse.getMessage());
-                mAwesomeValidation.validate();*/
                 showAlertMessage("" + deleteTokenResult.getMessage());
             }
             hideProgress();
         } catch (Exception e) {
             hideProgress();
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
         }
     }
 
     /**
-     * Implements a method to get MEMBER-ID from {@link android.content.SharedPreferences}
+     * Implements a method to get MEMBER-ID from {@link SharedPreferences}
      */
     public String getMemberId() {
         return loadPreferenceValue(ApplicationGlobal.KEY_MEMBERID, "10784");
     }
 
     /**
-     * Implements a method to get CLIENT-ID from {@link android.content.SharedPreferences}
+     * Implements a method to get CLIENT-ID from {@link SharedPreferences}
      */
     public String getClientId() {
         return loadPreferenceValue(ApplicationGlobal.KEY_CLUB_ID, ApplicationGlobal.TAG_CLIENT_ID);
@@ -636,11 +676,8 @@ public class DashboardActivity extends BaseActivity {
 
             @Override
             public void failure(RetrofitError error) {
-                //you can handle the errors here
                 Log.e(LOG_TAG, "RetrofitError : " + error);
                 hideProgress();
-
-                //getUnreadNewsCountService();
             }
         });
     }
@@ -651,7 +688,7 @@ public class DashboardActivity extends BaseActivity {
      */
     private void updateGetUnreadCountResponse(JsonObject jsonObject) {
 
-        Log.e(LOG_TAG, "SUCCESS RESULT : " + jsonObject.toString());
+        Log.e(LOG_TAG, "UNREAD NEWS COUNT RESPONSE : " + jsonObject.toString());
 
         Type type = new TypeToken<GetUnreadNewsResponse>() {
         }.getType();
@@ -669,17 +706,113 @@ public class DashboardActivity extends BaseActivity {
                     dashboardRecyclerAdapter.updateBadgerCount(getUnreadNewsResponse.getData().getUnRead());
                 }
             } else {
-                /*mAwesomeValidation.addValidation(etUserName, "regex", updatePasswordResponse.getMessage());
-                mAwesomeValidation.validate();*/
+                showAlertMessage("" + getUnreadNewsResponse.getMessage());
+            }
+        } catch (Exception e) {
+            Log.e(LOG_TAG, "" + e.getMessage());
+        }
+    }
+
+    /****************** ++ GET CLUB NEWS UNREAD COUNT FUNCTIONALITY ++ ******************/
+
+
+    /**
+     * Call Features flag web service to get list of
+     * features show on dashboard.
+     */
+    private void callFeaturesFlagService() {
+
+        showPleaseWait("Please wait...");
+
+        aJsonParamsFeaturesFlag = new AJsonParamsFeaturesFlag();
+        aJsonParamsFeaturesFlag.setCallid(ApplicationGlobal.TAG_GCLUB_CALL_ID);
+        aJsonParamsFeaturesFlag.setVersion(ApplicationGlobal.TAG_GCLUB_VERSION);
+
+        featureFlagsAPI = new FeatureFlagsAPI(getClientId(), "GETCLUBFEATURES", aJsonParamsFeaturesFlag, "CLUBINFO", ApplicationGlobal.TAG_GCLUB_MEMBERS);
+
+        //Creating a rest adapter
+        RestAdapter adapter = new RestAdapter.Builder()
+                .setEndpoint(WebAPI.API_BASE_URL)
+                .build();
+
+        //Creating an object of our api interface
+        WebServiceMethods api = adapter.create(WebServiceMethods.class);
+
+        api.getFeaturesFlagOptions(featureFlagsAPI, new Callback<JsonObject>() {
+            @Override
+            public void success(JsonObject jsonObject, retrofit.client.Response response) {
+
+                updateFeaturesFlagResponse(jsonObject);
+            }
+
+            @Override
+            public void failure(RetrofitError error) {
+                Log.e(LOG_TAG, "RetrofitError : " + error);
+                hideProgress();
+            }
+        });
+    }
+
+    /****************** ++ GET DASHBOARD FEATURES FLAG ++ ******************/
+
+    /**
+     * Implements a method which called to get count of
+     * UNREAD club news.
+     */
+    private void updateFeaturesFlagResponse(JsonObject jsonObject) {
+
+        Log.e(LOG_TAG, "FEATURES FLAG RESPONSE : " + jsonObject.toString());
+
+        Type type = new TypeToken<FeatureFlagsResponse>() {
+        }.getType();
+        featureFlagsResponse = new Gson().fromJson(jsonObject.toString(), type);
+
+        try {
+            /**
+             *  Check "Result" 1 or 0. If 1, means data received successfully.
+             */
+            if (featureFlagsResponse.getMessage().equalsIgnoreCase("Success")) {
+
+                //Make Dashboard dynamic according these bool values.
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_COURSE_DIARY_FEATURE, featureFlagsResponse.getData().getCourseDiaryFeatures());
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_COMPETITIONS_FEATURE, featureFlagsResponse.getData().getCompetitionsFeature());
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_HANDICAP_FEATURE, featureFlagsResponse.getData().getHandicapFeature());
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_MEMBERS_FEATURE, featureFlagsResponse.getData().getMembersFeature());
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_CLUB_NEWS_FEATURE, featureFlagsResponse.getData().getClubNewsFeature());
+                savePreferenceBooleanValue(ApplicationGlobal.KEY_YOUR_ACCOUNT_FEATURE, featureFlagsResponse.getData().getYourAccountFeature());
+
+                setGridMenuOptions();
+
+            } else {
                 showAlertMessage("" + getUnreadNewsResponse.getMessage());
             }
             hideProgress();
         } catch (Exception e) {
             hideProgress();
             Log.e(LOG_TAG, "" + e.getMessage());
-            e.printStackTrace();
         }
     }
 
-    /****************** ++ GET CLUB NEWS UNREAD COUNT FUNCTIONALITY ++ ******************/
+    /****************** ++ GET DASHBOARD FEATURES FLAG ++ ******************/
+
+    /**
+     * Check the device to make sure it has the Google Play Services APK. If
+     * it doesn't, display a dialog that allows users to download the APK from
+     * the Google Play Store or enable it in the device's system settings.
+     */
+    private boolean checkPlayServices() {
+        GoogleApiAvailability apiAvailability = GoogleApiAvailability.getInstance();
+        int resultCode = apiAvailability.isGooglePlayServicesAvailable(this);
+        if (resultCode != ConnectionResult.SUCCESS) {
+            if (apiAvailability.isUserResolvableError(resultCode)) {
+                apiAvailability.getErrorDialog(this, resultCode, PLAY_SERVICES_RESOLUTION_REQUEST)
+                        .show();
+            } else {
+                Log.i("checkPlayServices", "This device is not supported.");
+                finish();
+            }
+            return false;
+        }
+        return true;
+    }
 }