+++ Sunningdale/MHSystems/app/src/main/java/com/mh/systems/sunningdale/fragments/NewsWebCamFragment1.java
@@ -2,16 +2,24 @@ package com.mh.systems.sunningdale.fragments;
 
 
 import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
+import android.widget.ImageView;
 import android.widget.ProgressBar;
 
 import com.mh.systems.sunningdale.R;
+import com.mh.systems.sunningdale.activites.ClubNewsWebCamActivity;
+import com.mh.systems.sunningdale.adapter.RecyclerAdapter.ClubNewsSwipeAdapter;
+
+import java.io.InputStream;
 
 /**
  * The {@link NewsWebCamFragment1} used to display the detail of LOGIN
@@ -27,13 +35,13 @@ public class NewsWebCamFragment1 extends Fragment {
      *******************************/
     public final String LOG_TAG = NewsWebCamFragment2.class.getSimpleName();
 
-    String strURL = "http://www.sunningdale-golfclub.co.uk/visitor-information/webcam1/";
+//    String strURL = "http://www.sunningdale-golfclub.co.uk/visitor-information/webcam1/";
+      String strURL = "http://www.waidev2.com/~sunningdale-cam/camera1.jpg";
 
     /*********************************
      * INSTANCES OF CLASSES
      *******************************/
-    WebView wvNewsWebCam;
-    ProgressBar pbNewsCam;
+    ImageView ivNewsWebCam;
 
     private View mRootFragment;
 
@@ -41,41 +49,12 @@ public class NewsWebCamFragment1 extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         mRootFragment = inflater.inflate(R.layout.fragment_club_news_webcam, container, false);
 
-        //Initialize the view resources.
-        wvNewsWebCam = (WebView) mRootFragment.findViewById(R.id.wvNewsWebCam);
-        pbNewsCam = (ProgressBar) mRootFragment.findViewById(R.id.pbNewsCam);
+        ivNewsWebCam = (ImageView) mRootFragment.findViewById(R.id.ivNewsWebCam);
 
-        //Load Web View URL.
-        wvNewsWebCam.setWebViewClient(new myWebClient());
-        wvNewsWebCam.getSettings().setJavaScriptEnabled(true);
-        wvNewsWebCam.getSettings().setBuiltInZoomControls(true);
-        wvNewsWebCam.getSettings().setSupportZoom(true);
-        wvNewsWebCam.loadUrl(strURL);
+        // Image link from AWS server.
+        ((ClubNewsWebCamActivity)getActivity()).new DownloadImageFromInternet(ivNewsWebCam)
+                .execute(strURL);
 
         return mRootFragment;
-    }
-
-    /**
-     * Implements a class to Handle and display progress in
-     * Web View.
-     */
-    public class myWebClient extends WebViewClient {
-        @Override
-        public void onPageStarted(WebView view, String url, Bitmap favicon) {
-            super.onPageStarted(view, url, favicon);
-        }
-
-        @Override
-        public boolean shouldOverrideUrlLoading(WebView view, String url) {
-            pbNewsCam.setVisibility(View.VISIBLE);
-            view.loadUrl(url);
-            return true;
-        }
-
-        @Override
-        public void onPageFinished(WebView view, String url) {
-            super.onPageFinished(view, url);
-            pbNewsCam.setVisibility(View.GONE);
-        }
     }
 }